requirements.txt
flask
flask-cors
google-generativeai
pypdf
gunicorn
streamlit